Adding authorization for update and delete methods

+++ backend/controllers/goalController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require("express-async-handler");
 const goalSchema = require("../model/goalModel");
+const userSchema = require("../model/userModel");
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await goalSchema.find();
+  const goals = await goalSchema.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -11,6 +12,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await userSchema.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   await goalSchema.deleteOne({ id: req.params.id });
@@ -23,6 +36,19 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+
+  const user = await userSchema.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updatedGoal = await goalSchema.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -42,6 +68,7 @@ const createGoal = asyncHandler(async (req, res) => {
 
   const goal = await goalSchema.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
